schema.sql: INTEGER -> REAL

+++ other/schema.sql
@@ -1,12 +1,12 @@
 CREATE TABLE repos (reponame TEXT, PRIMARY KEY (reponame));
 CREATE TABLE packages (packagename TEXT, PRIMARY KEY (packagename));
 CREATE TABLE versioned_packages (
-	reponame TEXT, packagename TEXT, version TEXT, release INTEGER,
+	reponame TEXT, packagename TEXT, version TEXT, release REAL,
 	PRIMARY KEY (reponame, packagename));
 
 CREATE TABLE files (
 	reponame TEXT, packagename TEXT, version TEXT,
-	release INTEGER, filename TEXT,
+	release REAL, filename TEXT,
 	FOREIGN KEY (packagename) REFERENCES packages,
 	FOREIGN KEY (reponame) REFERENCES repos,
 	PRIMARY KEY (reponame, packagename, filename));
